+++ src/util/util.cpp
@@ -17,7 +17,7 @@ std::string util::to_string(const sf::Vector2f &vec) {
 }
 
 float util::getRandom(float min, float max) {
-	return min + static_cast<float>(fmodf(rand(), max + 1.0f));
+	return min + static_cast<float>(fmodf(static_cast<float>(rand()), max + 1.0f));
 }
 
 sf::Color util::toColor(float r, float g, float b) {
